requirements.txt
pandas
numpy
streamlit
streamlit-extras
streamlit-plotly-events
matplotlib
openai
boto3
umap-learn